+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.82 (not yet released)
 ------------
 
+* Added Action 'Other: Replace entity tokens in HTML'.
+
 Version 1.81
 ------------